Fixes [#3864702], problem with qtip showing up when hovering on student names
Basically javascript is stupid; at least jquery and qtip both behave
different when the environment has requirejs in it.

to solve this problem, we move the global requirejs functions to
window.ttm.

Also fixed a problem where main.js was overwriting the contents of
window.ttm instead of extending it.

ALSO this should fix the newly found silently failing tests on review